Polish both benchmarks
We are using iai and criterion with a custom measurement. The difference
between criterion and iai out of the box is supposed to be that one
measures wall-clock time (criterion) and the other uses instrumentation
to measure something correlated to wall-clock time.

However, out of the box criterion turned out to be too noisy to be
usable with this benchmark. It might be due to the function benchmarked
being too big.

Iai does not allow to separate setup, so it’s still nice to have
criterion. Plus criterion supports fancy stats out of the box, while iai
is supposed to be used with a cargo plugin.