doc: make sure the README is displayed on sr.ht
On sourcehut, the symlink is not followed for the main README.md file.
To have a rendered README on the project homepage on sourcehut, we need
to have a regular file at the top.

This commit fixes this and makes the nested README a symlink to the
top-level README. As far as I could test, this works nicely with `cargo
sync-readme`.