fix: adapt to rusqlite 0.29 and tighten the requirements
As discussed in #69, currently, version 1.0.1 of `rusqlite_migration` is
incompatible with rusqlite version 0.29. The problem is that the
Cargo.toml states the contrary, when it says:

```toml
[dependencies.rusqlite]
version = ">=0.23.0"
default-features = false
features = []
```

I think we should treat this as a bug and issue a patch release with a
fix.

This commit makes the code compatible with version 0.29. We are lucky
that such a code exists and is not too complicated.

As a result, and to fix the second related issue of insufficiently tight
version requirements on `rusqlite`, the range is not open-ended anymore
and we now constrain the upper bound as well. This means that we won’t
have to worry about breaking changes that `rusqlite` version 0.30 may
introduce.

I’ve tested this change with all the non-yanked versions of rusqlite and
it did `cargo build && cargo test tests::` without issues. The `cargo
test tests::` is to exclude documentation test as I don’t want to make
these weird due to the workaround for old `rusqlite` versions.
I’ve also run the full `cargo test` with `rusqlite` version 0.29 and
0.28, because the documentation examples should be correct for those.

Fixes #68