doc: make the deps badge more accurate
Currently, we report on the status of the dependencies in the dev
version (i.e. the lastest master commit of the GitHub repository). This
a bit misleading, because users of the library usually pull the latest
stable and use it. Whether dependencies are up to date for the latest
stable is quite different from the dev version.

This updated URL reports on the latest version. The drawback is that one
has to manually update the path when a new stable version is released.