doc: remove cargo-sync-readme
Use build.rs instead. This makes for a better contributing experience:
one just edits the README markdown file, instead of having to install
some external cargo tool. As a bonus, this change also simplifies CI.

The only drawback is that it makes compilation slightly longer for users
of the library but this really is very minimal (as far as I can measure
on my machine by changing build.rs and recompiling, less than 0.1s)

Note on the use of the write! macro: it uses write_all under the hood,
so it won’t silently discard part of the line if it doesn’t fit in the
buffer.