chore: fix cargo mutants for CI
From version 23.10, cargo mutants doesn’t run tests in different
packages. For us, this means that tests relying of features not enabled
by default and thus placed in different packages, are not executed. This
in turn causes a number of mutants to be deemed uncaught, but with no
mean to catch them since the tests associated with the code in different
features are not run.