Don't pin to exact rusqlite version (#120)
Pinning the dependency to the exact version `0.30.0` is unnecessary
because Cargo doesn't consider 0.30 and 0.31 compatible (Cargo's
modified semver considers the first non-zero component of the version as
the "major" number that determines compatibility). In addition, pinning
to an exact version can prevent an update to an in-theory-compatible
0.30.1 and even cause unresolvable dependency conflicts in downstream
packages. To be fair however, rusqlite doesn't seem to use the patch
component of the version number anyway so this wouldn't happen in
practice.