fix: failing examples with rusqlite “extra_check” feature
As reported in [1], in the documentation and examples, the following:

```conn.pragma_update(None, "journal_mode", &"WAL").unwrap();```

panics with `ExecuteReturnedResults` when the “extra_check” feature of
rusqlite is enabled.

This wasn’t caught automatically because it turns out our examples
aren’t run in CI, because the Cargo manifests weren’t properly
referenced.

This includes the manifests properly and fixes the CI.

[1] https://github.com/cljoly/rusqlite_migration/issues/15#issuecomment-1868524798