refactor: Keep migrations in `Cow` instead of `Vec`
This builds on https://github.com/cljoly/rusqlite_migration/pull/251 and
proposes another path. Like #251, it provides a method to store the
migrations in a `const` context, but:
* it does so without breaking changes, at least compared to the latest
  release: `new` is not const anymore, but that change was not released
  in a stable version.
* exposes only concrete types like `Vec` and slices to build the
  migrations, instead of the discussed `impl Into<Cow<'m, [M<'m>]>>`.
  This should make the compiler report clearer errors, improving
  usability.
* we can still easily built from a `Cow` type, by borrowing it.

Drawbacks
---------

The main drawback I can think of is this: the most intuitive function to
call (`new`) takes a `Vec`, while it would be better to just use a slice
(which `from_slice` takes). But that cost should be minimal and
preserves backward compatibility.

Co-authored-by: Joaquim Monteiro <joaquim.monteiro@protonmail.com>