test: improve build time by removing some unused optional deps
Looking at the report from `cargo t --timings`, `jiff` is one of the
longuest dependencies to build. It turns out that it is an optional
depenency of `env_logger`. This prompted me to go over a few of the
costliest dependencies that could be removed by opting out of default
dependencies.

* `env_logger`: we don’t really care about filtering, colored output,
  human times... most of the time.
* `simple-logging`: this is not used anymore and the feature of
  `env_logger` are useful sometimes, so no reason to go back to
  `simple-logging`.
* `criterion`: we rarely look at the benchmark output, we can enable the
  additional features temporarily only then.

Before the change, using `hyperfine -p 'cargo clean' 'cargo t'`:
```
Benchmark 1: cargo t
  Time (mean ± σ):     12.463 s ±  0.188 s    [User: 83.399 s, System: 9.080 s]
  Range (min … max):   12.256 s … 12.751 s    10 runs
```

After the change:
```
Benchmark 1: cargo t
  Time (mean ± σ):     10.821 s ±  0.110 s    [User: 62.565 s, System: 7.905 s]
  Range (min … max):   10.618 s … 10.951 s    10 runs
```

Gains of similar proportions are observed for test coverage and `cargo
mutants`.