test: add a test for the no leading 0 case
When defining migrations from a directory, the example pads the
migration id with leading zeros. While this is a good practice, if a
user doesn’t do that, we still accept the migrations and process them in
numerical order (which is not the lexicographic order, i.e. 1, 10, 2,
20, 3, 4...).

Add a test to capture both that this is valid and handled in numerical
order. This is both for completeness of the test suite and in case we
change the implementation (e.g. #288).