ci: introduce dependency cool down
Supply chain attacks have become quite common recently. To limit this
risk, [one simple idea](https://lobste.rs/s/rygog1/) is to have a
cool down period, where new dependencies are not merged. While I review
the changes made in a version, it’s possible I miss something and merge
a vulnerable version. So let’s be a bit more cautious moving forward.

Note that this only applies to updates that are not security fixes:

> The cooldown option is only available for version updates, not security updates.

[GitHub docs](https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference?versionId=free-pro-team%40latest&productId=code-security&restPage=dependabot#cooldown-)

While we are at it, remove the deps.rs badge. This is often broken,
displaying “depenencies unknown” and when it does work, it will say that
depenencies are out of date for longer because of the cooldown period.

Having a badge always green at the expense of a higher supply chain risk
is not the right tradeoff at this point.